cli: internal rename of create -> new
