+++ packages/cli/src/lib/new/FactoryRegistry.ts
