+++ packages/cli/src/lib/new/factories/frontendPlugin.test.ts
