+++ packages/cli/src/lib/new/factories/frontendPlugin.ts
