+++ packages/cli/src/lib/new/factories/pluginCommon.test.ts
