+++ packages/cli/src/lib/new/factories/pluginCommon.ts
