+++ packages/cli/src/lib/new/factories/scaffolderModule.test.ts
