+++ packages/cli/src/lib/new/factories/scaffolderModule.ts
